Loader-/Allokator-Leak-Cluster + latente Truncation gefixt (Bug-Hunt)
Echte Ressourcen-Lecks auf Loader-Fehler-/Edge-Pfaden (kein Fuzzer deckte sie ab; keine
Speichersicherheits-Verletzung -- reine Lecks/Truncations):

F1 (load_into_pd): ein ELF mit >MAX_IMG_SEGS(8) PT_LOAD-Segmenten wurde voll gemappt, aber nur
   die ersten 8 via loaded_register fuer den Teardown gemerkt -> Segmente 9+ leckten auch bei
   sauberem Laden+Abbau (Parser begrenzt phnum=u16 nicht). FIX: fail-closed, bevor etwas alloziert
   ist (img.segments().count() > MAX_IMG_SEGS -> cleanup).

F2 (load_into_pd): schlug eine Segment-/Stack-alloc, ein vspace_map_page_at oder der spawn
   mittendrin fehl, leckten die schon allozierten Daten-Frames (loaded_register laeuft erst am
   Ende; vspace_teardown sammelt nur Tabellen-Frames). FIX: `cleanup`-Closure gibt die bisher
   allozierten Segment-Frames (+ Stack) auf JEDEM Mid-Load-Fehlerpfad frei (Segmente werden vor
   dem Mappen in seglist registriert; MEM-Lock vor vspace_teardown freigegeben).

F3 (load_elf): erzeugte die PD und rief load_into_pd; bei dessen Fehler wurde der PD-Slot nie
   freigegeben (load_into_pd baut pd bewusst nicht ab -> Aufrufer-Pflicht). FIX: pds.free(pd) bei
   None (PD ist dann unbound + cap-frei).

M1 (PhysAllocator::alloc): checked_add-Overflow brach die GANZE Allokation ab statt das Fragment
   zu ueberspringen -> jetzt `continue`.
M2 (PhysAllocator::alloc): ein beidseitiger Verschnitt bei voller Free-Liste (MAX_FRAGMENTS) verlor
   das Suffix (remove + 2x insert, 2. insert scheitert) -> jetzt Fragment ueberspringen statt RAM
   still zu verlieren.
M3 (microkit Pd.partner): Option<u8> trunkierte bei NPDS>255 -> Option<u16> (zukunftssicher).

test-qemu == ALL PASS == (RC=0), Loader-L2/L3/L4 + Adversarial-Matrix gruen, keine Regression.
(M4/pcie 64-Bit-BAR war kein Bug: der size==0-Wrap-Check faengt BARs >=4 GiB bereits ab.)

+++ crates/sel4lake-microkit/src/lib.rs
@@ -260,7 +260,7 @@ struct Pd {
     /// Nur für [`Domain::HardwareLand`]: der **eine** unveränderliche Trusted-SAS-Partner
     /// (PD-Index), mit dem dieses Backend ausschließlich kommunizieren darf. Bei der Erzeugung
     /// gesetzt; `None` für Trusted/User.
-    partner: Option<u8>,
+    partner: Option<u16>,
     /// Stabile **Backend-Id** (ext-22): identifiziert ein HardwareLand-Backend unabhängig vom
     /// PD-Index (für später: mehrere NICs/USB-Controller, Hotplug, PCIe). `0` = keins.
     backend_id: u16,
@@ -379,7 +379,7 @@ impl PdTable {
         self.pds[i] = Pd {
             used: true,
             domain: Domain::HardwareLand,
-            partner: Some(partner as u8),
+            partner: Some(partner as u16),
             backend_id,
             chan_ep: ep,
             chan_ntfn: ntfn,
